+++ src/features/router/useSmoothRouter.tsx
@@ -1,8 +1,8 @@
 import { useContext, useCallback } from "react"
 
-import { PageStateContext } from "@src/contexts/PageStateContext"
 import { redirect } from "@src/features/router/"
 import { PageData, parsePageContent } from "@src/features/parser"
+import { PageStateContext } from "./PageStateContext"
 
 const useSmoothRouter = () => {
     const { setIsLoading, setPageData } = useContext(PageStateContext)
